Enhance similar images: hash size up to 8192, precise file sizes, improved ignore same size, size ratio filter
- Expand hash size options: 8, 16, 32, 64, 256, 512, 1024, 2048, 4096, 8192 (u8 -> u16)
- Add file size ratio filter (size_ratio_enabled, size_ratio) to exclude groups by max/min ratio
- Change ignore same size to discard entire group when all files have identical sizes
- Replace KiB/MiB formatting with precise KB/MB/GB display in similar images results
- Update all GUIs (GTK, Krokiet/Slint), CLI, and tests
- Update all dependencies

+++ krokiet/src/settings/model.rs
@@ -236,7 +236,7 @@ impl Default for SettingsCustom {
 
 pub struct ComboBoxItems {
     pub language: StringComboBoxItem<String>,
-    pub hash_size: StringComboBoxItem<u8>,
+    pub hash_size: StringComboBoxItem<u16>,
     pub resize_algorithm: StringComboBoxItem<FilterType>,
     pub image_hash_alg: StringComboBoxItem<HashAlg>,
     pub duplicates_hash_type: StringComboBoxItem<HashType>,
